wallet: option to only use pre/post fork fake outs for pre/post fork outs
If a pre-fork output is spent on both Monero and attack chain,
any post-fork output can be deduced to be a fake output, thereby
decreasing the effective ring size. This option, on by default,
allows selecting only pre-fork outputs in this case, so that the
same ring can be used when spending it on the other side, which
does not decrease the effective ring size.

This is intended to be SET when intending to spend Monero on the
attack fork, and to be UNSET if not intending to spend Monero
on the attack fork (since it leaks the fact that the output being
spent is pre-fork).

Similarly, if a post fork output is spent on its chain, and it
uses pre fork outputs, and if another post fork output is spent
on another chain, and also uses pre fork outputs, any pre fork
output in common can be deduced to be fake on at least one of
the forks, since if it was the real one on both, then it would
likely have the same key image.

If you use this option and intend to spend Monero on both forks,
then spend real Monero first.