Allow up to 10% of txs in a block to be of ringsize 1 while disallowing ringsize 2
The basic idea behind the block template filling algorithm is as follows:

The algorithm consists of two passes, the first for the normal txes (i.e.
ringsize >= 3) and the second for ringsize-1 txes. In the first pass, normal txes are
added as much as possible to the block as before, while ignoring ringsize-1 txes. Since
txes in m_txs_by_fee_and_receive_time are sorted, the txes stored in bl.tx_hashes will
be also sorted, with the last one being the least preferrable.

In the second pass, we go through txes in m_txs_by_fee_and_receive_time again, this time
only looking at ringsize-1 txes. For each ringsize-1 tx, we decide if it should be taken
or not by considering the following two possibilities:

  1) The block still has some free space for the ringsize-1 tx to be added.

  2) The above isn't the case (the block is already pretty full), but the ringsize-1 tx
     can still be added by replacing an existing normal tx that is the least preferrable
     which is stored at the back of bl.tx_hashes.

In both cases, the ringsize-1 tx can be added only if:

    M <= ceil(N * NOFAKE_TXS_TO_TOTAL_TXS_PERCENT / 100)

where M and N are the number of ringsize-1 txes and the number of all txes in the
block, respectively.